backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pillow==10.3.0
opencv-python==4.10.0.84
numpy==1.26.4
pytesseract==0.3.10
reportlab==4.2.2
